refactor: remove facade/collection
