+++ modules/@angular/core/src/change_detection/change_detection_util.ts
@@ -6,8 +6,7 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-import {areIterablesEqual, isListLikeIterable} from '../facade/collection';
-import {isPrimitive, looseIdentical} from '../facade/lang';
+import {getSymbolIterator, isJsObject, isPrimitive, looseIdentical} from '../facade/lang';
 
 export {looseIdentical} from '../facade/lang';
 
@@ -77,4 +76,39 @@ export class SimpleChange {
    * Check whether the new value is the first value assigned.
    */
   isFirstChange(): boolean { return this.firstChange; }
+}
+
+export function isListLikeIterable(obj: any): boolean {
+  if (!isJsObject(obj)) return false;
+  return Array.isArray(obj) ||
+      (!(obj instanceof Map) &&      // JS Map are iterables but return entries as [k, v]
+       getSymbolIterator() in obj);  // JS Iterable have a Symbol.iterator prop
+}
+
+export function areIterablesEqual(
+    a: any, b: any, comparator: (a: any, b: any) => boolean): boolean {
+  const iterator1 = a[getSymbolIterator()]();
+  const iterator2 = b[getSymbolIterator()]();
+
+  while (true) {
+    const item1 = iterator1.next();
+    const item2 = iterator2.next();
+    if (item1.done && item2.done) return true;
+    if (item1.done || item2.done) return false;
+    if (!comparator(item1.value, item2.value)) return false;
+  }
+}
+
+export function iterateListLike(obj: any, fn: (p: any) => any) {
+  if (Array.isArray(obj)) {
+    for (let i = 0; i < obj.length; i++) {
+      fn(obj[i]);
+    }
+  } else {
+    const iterator = obj[getSymbolIterator()]();
+    let item: any;
+    while (!((item = iterator.next()).done)) {
+      fn(item.value);
+    }
+  }
 }
