+++ modules/@angular/core/src/util/lang.ts
@@ -24,3 +24,18 @@ export function isPromise(obj: any): obj is Promise<any> {
 export function isObservable(obj: any | Observable<any>): obj is Observable<any> {
   return !!(obj && obj[symbolObservable]);
 }
+
+// TODO(misko): replace with Object.assign once we require ES6.
+export function merge<V>(m1: {[key: string]: V}, m2: {[key: string]: V}): {[key: string]: V} {
+  const m: {[key: string]: V} = {};
+
+  for (const k of Object.keys(m1)) {
+    m[k] = m1[k];
+  }
+
+  for (const k of Object.keys(m2)) {
+    m[k] = m2[k];
+  }
+
+  return m;
+}
