+++ modules/@angular/platform-browser/src/dom/debug/by.ts
@@ -6,10 +6,8 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-import {DebugElement, Type} from '@angular/core';
-
+import {DebugElement, Predicate, Type} from '@angular/core';
 import {getDOM} from '../../dom/dom_adapter';
-import {Predicate} from '../../facade/collection';
 import {isPresent} from '../../facade/lang';
